pyproject.toml
[project]
name = "release-tools"
version = "0.1.0"
description = "Python classes for GitHub Actions release workflows"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gitpython",
    "invoke",
    "python-dotenv",
    "pygithub",
    "semver",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/release_tools"]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "COM812",  # conflicts with ruff formatter
    "D203",  # incompatible with D211
    "D213",  # incompatible with D212
    "T201",  # print is used intentionally
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "ruff",
    "ty",
]
